Clean `:code-gen` module + update its README
